pyproject.toml
[tool.poetry]
name = "FlashCardSRSWeb"
version = "0.1.0"
description = "High customizability and low friction system to memorize anything!"
authors = ["José Lima <josesblima@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.9,<0.116.0"
sqlalchemy = ">=2.0.38,<3.0.0"
uvicorn = ">=0.34.0,<0.35.0"
pydantic = ">=2.12,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
psycopg = {version = "^3.2", extras = ["binary"]}
alembic = ">=1.14.1,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
pytest-asyncio = "^0.25.3"
inject = "^5.2.1"
py-automapper = "^2.2.0"
greenlet = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.4,<9.0.0"
pytest-asyncio = ">=0.25.3,<0.26.0" 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportCallIssue = "none"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
